travel_planner_fase3/travel_planner/requirements.txt
streamlit==1.32.0
sqlalchemy==2.0.29
plotly==5.20.0
pandas==2.2.1
python-dotenv==1.0.1
Pillow==10.2.0
amadeus==9.0.0
anthropic>=0.91.0
